Exclude tests which make network calls
