+++ cmd/core/cmd_test.go
@@ -1,3 +1,5 @@
+// +build integration
+
 package cmdcore
 
 import (
