+++ cmd/sandbox/start_test.go
@@ -1,3 +1,5 @@
+// +build integration
+
 package sandbox
 
 import (
