+++ cmd/version/version_test.go
@@ -1,3 +1,5 @@
+// +build integration
+
 package version
 
 import (
