+++ pkg/util/util_test.go
@@ -29,24 +29,6 @@ func TestPrintSandboxMessage(t *testing.T) {
 	})
 }
 
-func TestSendRequest(t *testing.T) {
-	t.Run("Successful get request", func(t *testing.T) {
-		response, err := SendRequest("GET", "https://github.com", nil)
-		assert.Nil(t, err)
-		assert.NotNil(t, response)
-	})
-	t.Run("Successful get request failed", func(t *testing.T) {
-		response, err := SendRequest("GET", "htp://github.com", nil)
-		assert.NotNil(t, err)
-		assert.Nil(t, response)
-	})
-	t.Run("Successful get request failed", func(t *testing.T) {
-		response, err := SendRequest("GET", "https://github.com/evalsocket/flyte/archive/refs/tags/source-code.zip", nil)
-		assert.NotNil(t, err)
-		assert.Nil(t, response)
-	})
-}
-
 func TestIsVersionGreaterThan(t *testing.T) {
 	t.Run("Compare FlyteCTL version when upgrade available", func(t *testing.T) {
 		_, err := IsVersionGreaterThan("v1.1.21", testVersion)
